avoid a legitimate -Wstrict-overflow warning
* libidu/scanners.h (log_8_member_files): Change type from int to size_t.
* libidu/scanners.c: Likewise.

+++ libidu/scanners.c
@@ -40,7 +40,7 @@ static struct obstack lang_args_obstack;
 struct lang_args *lang_args_default = 0;
 struct lang_args *lang_args_list = 0;
 struct obstack tokens_obstack;
-int log_8_member_files = 0;
+size_t log_8_member_files = 0;
 
 extern void usage (void);
 extern char *program_name;
